pyproject.toml
[tool.poetry]
name = "django-live-dashboard"
version = "0.1.0"
description = ""
authors = ["Adam Hill <adamghill@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.6"
hiredis = "^1.0.1"
aioredis = "^1.3.1"
django = "^3.0.5"
orjson = "^2.6.5"
django-app-settings = "^0.7.0"
python-box = "^4.2.2"
redis = "^3.4.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
flake8 = "^3.7.9"
black = "^19.10b0"
uvicorn = "^0.11.3"
hiredis = "^1.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
